Move ACOptions back to autohint.py
autohint.hintFiles take ACOptions as argument, but having ACOptions not
defined in autohint.py means every caller needs to define it by itself,
which is not good. Move it back, and add a subclass in __main__.py that
takes parsed args as an argument.
